+++ CHANGES.md
@@ -8,6 +8,8 @@ Version 0.7.1 (Pending)
 - More attendee validation for at-con orders
 - Fixed delete button for attendees in at-con orders
 - Only show printer/till selections to users that can use them
+- Use regular input field for birthdates (so mobile devices don't pop up calendars and make you scroll through
+  every month to pick a date)
 
 Version 0.7.0 (6/9/2019)
 - When creating database, add citext extension if it does not exist (needed to support case insensitive searches,